refactor(coins): move coin-dto mapper to  mapper package
