+++ remote/src/main/java/com/hxl/remote/CoinRemoteImpl.java
@@ -2,9 +2,10 @@ package com.hxl.remote;
 
 import com.hxl.data.repository.coin.CoinRemote;
 import com.hxl.domain.model.Coin;
+import com.hxl.domain.model.History;
 import com.hxl.remote.api.CoinService;
 import com.hxl.remote.model.CoinDTO;
-import com.hxl.remote.model.CoinDTOMapper;
+import com.hxl.remote.mapper.CoinDTOMapper;
 import com.hxl.remote.model.Response;
 
 import java.util.List;
